requirements.txt
langchain==0.2.5
langchain-core>=0.2.7,<0.3.0
langchain-community>=0.2.4,<0.3.0
langchain-text-splitters>=0.2.1,<0.3.0
langchain-huggingface>=0.0.3
langchain-groq>=0.1.5

faiss-cpu
unstructured

fastapi
uvicorn
pydantic
requests

python-dotenv 

streamlit